requirements.txt
PyPDF2
anthropic
gspread
google-auth
